# src/main/resources/api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Football team management",
    "version": "3.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8888",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "SecurityScheme": []
    }
  ],
  "paths": {
...
    "/trainings/{training}": {
      "put": {
        "tags": [
          "training-controller"
        ],
        "operationId": "update_1",
        "parameters": [
          {
            "name": "training",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseTrainingResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "training-controller"
        ],
        "operationId": "delete_1",
        "parameters": [
          {
            "name": "training",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    },
    "/teams/{team}": {
      "get": {
        "tags": [
          "team-controller"
        ],
        "operationId": "getTeam",
        "parameters": [
          {
            "name": "team",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseTeamResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "team-controller"
        ],
        "operationId": "update_2",
        "parameters": [
          {
            "name": "team",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeamRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseTeamResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "team-controller"
        ],
        "operationId": "delete_2",
        "parameters": [
          {
            "name": "team",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    },
...
    "/users/{userId}/upload-avatar": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "operationId": "uploadUserAvatar",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "avatar"
                ],
                "type": "object",
                "properties": {
                  "avatar": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseUserResponse"
                }
              }
            }
          }
        }
      }
    },
    "/trainings": {
      "get": {
        "tags": [
          "training-controller"
        ],
        "operationId": "getAll_1",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListTrainingResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "training-controller"
        ],
        "operationId": "create_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseTrainingResponse"
                }
              }
            }
          }
        }
      }
    },
    "/teams": {
      "get": {
        "tags": [
          "team-controller"
        ],
        "operationId": "getAll_2",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListTeamResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "team-controller"
        ],
        "operationId": "create_2",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeamRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseTeamResponse"
                }
              }
            }
          }
        }
      }
    },
    "/teams/{teamId}/upload-logo": {
      "post": {
        "tags": [
          "team-controller"
        ],
        "operationId": "uploadUserAvatar_1",
        "parameters": [
          {
            "name": "teamId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "logo"
                ],
                "type": "object",
                "properties": {
                  "logo": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseTeamResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/votes/advance-search-with-specification": {
      "get": {
        "tags": [
          "vote-controller"
        ],
        "operationId": "advanceSearchWithSpecifications",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          },
          {
            "name": "vote",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResponseListVoteResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/users/advance-search-with-specification": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "operationId": "advanceSearchWithSpecifications_1",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          },
          {
            "name": "user",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResponseListUserResponse"
                }
              }
            }
          }
        }
      }
    },
    "/trainings/advance-search-with-specification": {
      "get": {
        "tags": [
          "training-controller"
        ],
        "operationId": "advanceSearchWithSpecifications_2",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          },
          {
            "name": "training",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResponseListTrainingResponse"
                }
              }
            }
          }
        }
      }
    },
    "/teams/advance-search-with-specification": {
      "get": {
        "tags": [
          "team-controller"
        ],
        "operationId": "advanceSearchWithSpecifications_3",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          },
          {
            "name": "team",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResponseListTeamResponse"
                }
              }
            }
          }
        }
      }
    },
    "/roles/advance-search-with-specification": {
      "get": {
        "tags": [
          "role-controller"
        ],
        "operationId": "advanceSearchWithSpecifications_4",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          },
          {
            "name": "role",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResponseListRoleResponse"
                }
              }
            }
          }
        }
      }
    },
    "/permissions/advance-search-with-specification": {
      "get": {
        "tags": [
          "permission-controller"
        ],
        "operationId": "advanceSearchWithSpecifications_5",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          },
          {
            "name": "permission",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResponseListPermissionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/matches/advance-search-with-specification": {
      "get": {
        "tags": [
          "match-controller"
        ],
        "operationId": "advanceSearchWithSpecifications_6",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          },
          {
            "name": "match",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResponseListMatchResponse"
                }
              }
            }
          }
        }
      }
    },
    "/eventParticipants/advance-search-with-specification": {
      "get": {
        "tags": [
          "event-participant-controller"
        ],
        "operationId": "advanceSearchWithSpecifications_7",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          },
          {
            "name": "participant",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResponseListEventParticipantResponse"
                }
              }
            }
          }
        }
      }
    },
    "/contents/advance-search-with-specification": {
      "get": {
        "tags": [
          "content-controller"
        ],
        "operationId": "advanceSearchWithSpecifications_8",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          },
          {
            "name": "content",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResponseListContentResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponseUserResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "result": {
            "$ref": "#/components/schemas/UserResponse"
          }
        }
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "dob": {
            "type": "string",
            "format": "date"
          },
          "roleIds": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "avatar": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "EventParticipantRequest": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "eventId": {
            "type": "string"
          },
          "roleInEvent": {
            "type": "string",
            "enum": [
              "COACH",
              "CAPTAIN",
              "PLAYER",
              "MEDICINE_TEAM",
              "OTHER"
            ]
          }
        }
      },
      "TrainingRequest": {
        "type": "object",
        "properties": {
          "teamId": {
            "type": "string"
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "participants": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventParticipantRequest"
            }
          }
        }
      },
      "ApiResponseTrainingResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "result": {
            "$ref": "#/components/schemas/TrainingResponse"
          }
        }
      },
...
      "TrainingResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "teamId": {
            "type": "string"
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "participants": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventParticipantResponse"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "MemberRequest": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "TeamRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "creator": {
            "type": "string"
          },
          "members": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MemberRequest"
            }
          }
        }
      },
      "ApiResponseTeamResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "result": {
            "$ref": "#/components/schemas/TeamResponse"
          }
        }
      },
...
      "TeamResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "creatorId": {
            "type": "string"
          },
          "teamMembers": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MemberResponse"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "logo": {
            "type": "string"
          }
        }
      },
...
      "ApiResponseVoid": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "result": {
            "type": "object"
          }
        }
      },
...
      "Pageable": {
        "type": "object",
        "properties": {
          "page": {
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "ApiResponsePageResponseListVoteResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "result": {
            "$ref": "#/components/schemas/PageResponseListVoteResponse"
          }
        }
      },
      "PageResponseListVoteResponse": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalPage": {
            "type": "integer",
            "format": "int64"
          },
          "totalItem": {
            "type": "integer",
            "format": "int64"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VoteResponse"
            }
          }
        }
      },
...
      "ApiResponsePageResponseListUserResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "result": {
            "$ref": "#/components/schemas/PageResponseListUserResponse"
          }
        }
      },
      "PageResponseListUserResponse": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalPage": {
            "type": "integer",
            "format": "int64"
          },
          "totalItem": {
            "type": "integer",
            "format": "int64"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserResponse"
            }
          }
        }
      },
      "ApiResponseListTrainingResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "result": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TrainingResponse"
            }
          }
        }
      },
      "ApiResponsePageResponseListTrainingResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "result": {
            "$ref": "#/components/schemas/PageResponseListTrainingResponse"
          }
        }
      },
      "PageResponseListTrainingResponse": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalPage": {
            "type": "integer",
            "format": "int64"
          },
          "totalItem": {
            "type": "integer",
            "format": "int64"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TrainingResponse"
            }
          }
        }
      },
      "ApiResponseListTeamResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "result": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TeamResponse"
            }
          }
        }
      },
      "ApiResponsePageResponseListTeamResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "result": {
            "$ref": "#/components/schemas/PageResponseListTeamResponse"
          }
        }
      },
...
      "ApiResponsePageResponseListRoleResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "result": {
            "$ref": "#/components/schemas/PageResponseListRoleResponse"
          }
        }
      },
...
      "ApiResponsePageResponseListPermissionResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "result": {
            "$ref": "#/components/schemas/PageResponseListPermissionResponse"
          }
        }
      },
...
      "ApiResponsePageResponseListMatchResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "result": {
            "$ref": "#/components/schemas/PageResponseListMatchResponse"
          }
        }
      },
...
      "ApiResponsePageResponseListEventParticipantResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "result": {
            "$ref": "#/components/schemas/PageResponseListEventParticipantResponse"
          }
        }
      },
...
      "ApiResponsePageResponseListContentResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "result": {
            "$ref": "#/components/schemas/PageResponseListContentResponse"
          }
        }
      },
...
    },
    "securitySchemes": {
      "SecurityScheme": {
        "type": "http",
        "name": "SecurityScheme",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}